--- a/plugins/MultiSites/.gitignore
+++ b/plugins/MultiSites/.gitignore
@@ -1,0 +1,1 @@
+tests/System/processed/*xml