… CHANGELOG.md
… core/DataAccess/ArchiveSelector.php
… core/DataTable.php
… core/DataTable/Filter/ColumnCallbackAddMetadata.php
… core/DataTable/Manager.php
… core/Site.php
… lang/en.json
… plugins/MultiSites/API.php
… plugins/MultiSites/Columns/Metrics/EcommerceOnlyEvolutionMetric.php
… plugins/MultiSites/Controller.php
… plugins/MultiSites/Dashboard.php
… plugins/MultiSites/MultiSites.php
… plugins/MultiSites/angularjs/dashboard/dashboard-group.filter.js
… plugins/MultiSites/angularjs/dashboard/dashboard-model.service.js
… plugins/MultiSites/angularjs/dashboard/dashboard.controller.js
… plugins/MultiSites/angularjs/dashboard/dashboard.directive.html
… plugins/MultiSites/angularjs/dashboard/dashboard.directive.less
… plugins/MultiSites/tests/Fixtures/ManySitesWithVisits.php
… plugins/MultiSites/tests/Integration/ControllerTest.php
… plugins/MultiSites/tests/Integration/DashboardTest.php
… tests/PHPUnit/Integration/Archive/DataTableFactoryTest.php
… tests/PHPUnit/System/expected/test_BackwardsCompatibility1XTest__MultiSites.getAll_day.xml
… tests/PHPUnit/System/expected/test_ImportLogs__MultiSites.getAll_month.xml
… tests/PHPUnit/System/expected/test_ImportLogs_withEnhancedAndLast7__MultiSites.getAll_month.xml
… tests/PHPUnit/System/expected/test_OneVisitorTwoVisits__MultiSites.getAll_day.xml
… tests/PHPUnit/System/expected/test_TwoVisitors_twoWebsites_differentDays_Conversions_MultiSites.getAll_firstSite_lastN__API.getProcessedReport_day.xml
… tests/PHPUnit/System/expected/test_TwoVisitors_twoWebsites_differentDays_Conversions__MultiSites.getAll_day.xml
… tests/PHPUnit/System/expected/test_TwoVisitors_twoWebsites_differentDays_Conversions__MultiSites.getAll_month.xml
… tests/PHPUnit/System/expected/test_TwoVisitors_twoWebsites_differentDays_Conversions__MultiSites.getAll_week.xml
… tests/PHPUnit/System/expected/test_TwoVisitors_twoWebsites_differentDays_Conversions__MultiSites.getAll_year.xml