@mattab 2015-04-22
Needs a rebase & then +1 to merge it before the beta/RC